+++ css/dubplus.css
@@ -389,7 +389,7 @@
   background-image: url(http://i.imgur.com/oTfAI0O.png); }
 
 .tasty-previews {
-  background-image: url(https://raw.githubusercontent.com/FranciscoG/DubPlus/no-gulp/emotes/tastycat/21.png); }
+  background-image: url(https://raw.githubusercontent.com/FranciscoG/DubPlus/undefined/emotes/tastycat/21.png); }
 
 #autocomplete-preview span {
   font-size: 0.8em;
