+++ src/js/dubplus.js
@@ -11,7 +11,7 @@
                                             
     https://github.com/DubPlus/DubPlus
 
-    MIT License
+    MIT License 
 
     Copyright (c) 2017 DubPlus
